extract api building blocks into api-common
