+++ packages/api-common/.gitattributes
@@ -1,0 +1,7 @@
+# Path-based git attributes
+# https://www.kernel.org/pub/software/scm/git/docs/gitattributes.html
+
+# Ignore all test and documentation with "export-ignore".
+/.gitattributes   export-ignore
+/.gitignore       export-ignore
+/README.md        export-ignore
