+++ packages/api-common/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@mainsail/api-http",
+	"name": "@mainsail/api-common",
 	"version": "0.0.1",
-	"description": "HTTP API for Mainsail",
+	"description": "Common API building blocks for Mainsail",
 	"license": "MIT",
 	"contributors": [],
 	"main": "distribution/index.js",
@@ -16,21 +16,17 @@
 		"test": "uvu -r tsm source .test.ts",
 		"test:coverage": "c8 pnpm run test",
 		"test:coverage:html": "c8 -r html --all pnpm run test",
-		"test:file": "uvu -r tsm source",
-		"test:integration": "uvu -r tsm integration .test.ts"
+		"test:file": "uvu -r tsm source"
 	},
 	"dependencies": {
 		"@hapi/boom": "9.1.4",
 		"@hapi/hapi": "20.1.5",
 		"@hapi/hoek": "9.2.0",
-		"@mainsail/api-common": "workspace:*",
 		"@mainsail/api-database": "workspace:*",
 		"@mainsail/container": "workspace:*",
 		"@mainsail/contracts": "workspace:*",
 		"@mainsail/kernel": "workspace:*",
-		"@mainsail/transactions": "workspace:*",
 		"@mainsail/utils": "workspace:*",
-		"dayjs": "1.10.7",
 		"joi": "17.9.2",
 		"nanomatch": "1.2.13",
 		"rate-limiter-flexible": "1.3.2",
@@ -42,15 +38,9 @@
 		"@types/hapi__joi": "17.1.7",
 		"@types/ip": "1.1.0",
 		"@types/semver": "6.2.3",
-		"got": "11.8.6",
 		"uvu": "^0.5.6"
 	},
 	"engines": {
 		"node": ">=20.x"
-	},
-	"arkecosystem": {
-		"core": {
-			"alias": "api-http"
-		}
 	}
 }
