--- packages/api-http/source/types/resources.ts
@@ -1,9 +1,0 @@
-import { BigNumber } from "@mainsail/utils";
-
-export type WalletResource = {
-	address: string;
-	publicKey?: string;
-	balance: BigNumber;
-	nonce: BigNumber;
-	attributes: object;
-};
