+++ packages/core/package.json
@@ -27,6 +27,7 @@
 	"dependencies": {
 		"@mainsail/api-database": "workspace:*",
 		"@mainsail/api-development": "workspace:*",
+		"@mainsail/api-common": "workspace:*",
 		"@mainsail/api-http": "workspace:*",
 		"@mainsail/api-sync": "workspace:*",
 		"@mainsail/bootstrap": "workspace:*",
